Fix: Changed the GetNotApprovedApplicationsInAProject to be located in projoect and let the endpoint take in the id or something
