+++ Services/Applications/IApplicationService.cs
@@ -5,6 +5,5 @@ namespace Lagalt_Backend.Services.ApplicationServices
     public interface IApplicationService : ICrudService<Application, int>
     {
         public Task<bool> ApplicationExists(int id);
-        public Task<ICollection<Application>> GetNotApprovedApplications(int projectId);
     }
 }
