+++ Services/Projects/IProjectService.cs
@@ -8,5 +8,6 @@ namespace Lagalt_Backend.Services.Projects
     public interface IProjectService : ICrudService<Project, int>
     {
         public Task<ICollection<Project>> GetProjectsBySkill(string skill);
+        public Task<ICollection<Application>> GetNotApprovedApplications(int projectId);
     }
 }
